% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_payload.R
\name{.build_payload}
\alias{.build_payload}
\title{Internal: temporary dummy payload builder}
\usage{
.build_payload(disease, geography, year, stratify_by = NULL)
}
\description{
Internal: temporary dummy payload builder
}
\keyword{internal}
