% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-metadata.R
\name{.decode_atlas}
\alias{.decode_atlas}
\title{Internal: decode fields returned by AtlasPlus}
\usage{
.decode_atlas(df)
}
\description{
Internal: decode fields returned by AtlasPlus
}
\keyword{internal}
