% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-metadata.R
\name{.decode_var}
\alias{.decode_var}
\title{Internal: decode codes using varvals table}
\usage{
.decode_var(ids, type, varvals)
}
\description{
Internal: decode codes using varvals table
}
\keyword{internal}
