% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-metadata.R
\name{.fetch_metadata}
\alias{.fetch_metadata}
\title{Internal: fetch AtlasPlus metadata from getInitData/00}
\usage{
.fetch_metadata(ttl = NULL)
}
\description{
Internal: fetch AtlasPlus metadata from getInitData/00
}
\keyword{internal}
