% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-request.R
\name{.request_atlas}
\alias{.request_atlas}
\title{Internal AtlasPlus request helper with caching}
\usage{
.request_atlas(payload)
}
\description{
Cached wrapper around \code{.request_atlas_uncached()}.
}
\keyword{internal}
