% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{.request_json}
\alias{.request_json}
\title{Internal HTTP request helper for CDCAtlas}
\usage{
.request_json(path, method = c("GET", "POST"), body = NULL, ttl = NULL)
}
\description{
Internal HTTP request helper for CDCAtlas
}
\keyword{internal}
