% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_atlas.R
\name{get_atlas}
\alias{get_atlas}
\title{Retrieve data from CDC AtlasPlus}
\usage{
get_atlas(disease, geography, year, stratify_by = NULL)
}
\arguments{
\item{disease}{Character scalar, e.g. "Gonorrhea"}

\item{geography}{"state", "county", "national"}

\item{year}{Numeric vector of years, e.g. 2018:2022.  Valid values are 2000-2023.}

\item{stratify_by}{Optional vector of variables to stratify by.  Options are "age", "race", "sex"}

\item{restrict_to}{Optional vector of states to limit geography}
}
\value{
A tibble
}
\description{
Retrieve data from CDC AtlasPlus
}
